backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altgraph==0.17.3
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
auto-py-to-exe==2.37.0
autopep8==2.0.1
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.34.54
botocore==1.34.54
bottle==0.12.25
bottle-websocket==0.2.9
CacheControl==0.14.2
cachetools==5.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
codetiming==1.4.0
colorama==0.4.5
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.26.3
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-debug-toolbar==4.4.2
djangorestframework==3.15.2
Eel==0.16.0
environs==11.0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.1.0
fastapi==0.110.0
fastapi_cors==0.0.6
fastjsonschema==2.21.1
feedparser==6.0.10
filelock==3.13.1
firebase-admin==6.7.0
Flask==2.1.3
flask-cors==5.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==2.0.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.21.2
idna==3.4
importlib-metadata==7.0.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langsmith==0.2.10
MarkupSafe==2.1.1
marshmallow==3.22.0
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mysqlclient==2.2.4
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.4
openai==1.59.4
openpyxl==3.1.2
orjson==3.10.13
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.2.0
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.1.2
Pygments==2.19.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.10.1
pyparsing==3.1.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pywin32==308
pywin32-ctypes==0.2.2
pywinpty==2.0.14
PyYAML==6.0.1
pyzmq==26.2.0
realpython-reader==1.1.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.39
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.38.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.27.1
virtualenv==20.26.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.1.2
whichcraft==0.6.1
yarl==1.18.3
zipp==3.17.0
zope.event==5.0
zope.interface==6.0
